fix: corregi la estructura de carpetas y configure el module-info.java para que los tests puedan acceder al modelo
