+++ src/main/test/java/edu/fiuba/algo3/modelo/BloqueLapizAbajoTest.java
@@ -1,0 +1,20 @@
+package edu.fiuba.algo3.modelo;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class BloqueLapizAbajoTest {
+
+    @Test
+    public void test01CreoUnPersonajeConElLapizArribaYBajoSuLapizUsandoElBloqueLapizAbajo(){
+
+        Personaje personaje = new Personaje();
+        BloqueLapizAbajo bloqueLapizAbajo = new BloqueLapizAbajo();
+
+        bloqueLapizAbajo.ejecutar(personaje);
+
+
+        assertEquals(LapizAbajo.class,personaje.direccionLapiz().getClass());
+    }
+}
